# api/docs/swagger.yaml
definitions:
...
  memory.GetAllRes:
    properties:
      count:
        type: integer
      memories:
        items:
          $ref: '#/definitions/memory.MemoryRes'
        type: array
    type: object
...
  memory.MemoryCreateBody:
    properties:
      date:
        type: string
      description:
        type: string
      locations:
        $ref: '#/definitions/memory.Point'
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  memory.MemoryRes:
    properties:
      date:
        type: string
      description:
        type: string
      locations:
        $ref: '#/definitions/memory.Point'
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  memory.Point:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for Time Capsule resources
  title: Time Capsule API
  version: "1.0"
paths:
...
  /memories:
    delete:
      consumes:
      - application/json
      description: Delete a memory by ID
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Memory deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a memory
      tags:
      - memories
    get:
      consumes:
      - application/json
      description: Get a memory by ID
      parameters:
      - description: Memory ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.MemoryRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a memory
      tags:
      - memories
    post:
      consumes:
      - application/json
      description: Create a new memory
      parameters:
      - description: MemoryCreateBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/memory.MemoryCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Memory created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a memory
      tags:
      - memories
    put:
      consumes:
      - application/json
      description: Update a memory by ID
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      - description: MemoryCreateBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/memory.MemoryCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Memory updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a memory
      tags:
      - memories
  /memories/all:
    get:
      consumes:
      - application/json
      description: Get all memories with optional filters
      parameters:
      - description: Start Date
        in: query
        name: start_date
        type: string
      - description: End Date
        in: query
        name: end_date
        type: string
      - description: Tag
        in: query
        name: tag
        type: string
      - description: Type
        in: query
        name: type
        type: string
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.GetAllRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all memories
      tags:
      - memories
  /memories/hictorical:
    get:
      consumes:
      - application/json
      description: Get all historical memories with optional filters
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.GetAllRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all historical memories
      tags:
      - memories
  /memories/others:
    get:
      consumes:
      - application/json
      description: Get memories of others with optional filters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.GetAllRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get memories of others
      tags:
      - memories
  /memories/tag:
    get:
      consumes:
      - application/json
      description: Get by tag memories with optional filters
      parameters:
      - description: Tag
        in: query
        name: tag
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.GetAllRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get by tag memories
      tags:
      - memories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"